requirements.txt
## Code quality
pytest>=7.0.1
black==22.3.0

## Standard science stack
h5py
hdf5plugin
matplotlib
numpy>=1.25.0,<=1.26.4 # numpy>=1.25.0 is needed for pytorch-optimizer, numpy<=1.26.4 is needed for numba

## Standard utils
tqdm
wget

## Deep learning
einops
opt_einsum
torch>=2.6.0
torch_geometric>=2.4.0
torch-ema
pytorch_optimizer>=3.2.0

## Experiment management
hydra-core
mlflow==2.4.1 # avoid database version issues
optuna

## Project-specific: general
clifford>=1.4.0
numba==0.58.1 # clifford uses old numba code
xformers # there exists known compatibility issues with xformers on Mac systems
torchdiffeq # for CFM sampling

energyflow
awkward
lgatr